third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/createDocument-doctype-err.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var doctype = document.implementation.createDocumentType("x", "x", "x");
shouldBeType("doctype", "DocumentType");
shouldNotThrow("document.implementation.createDocument(null, null, doctype)");
shouldNotThrow("document.implementation.createDocument(null, null, undefined)");
shouldNotThrow("document.implementation.createDocument(null, null, null)");

shouldThrow("document.implementation.createDocument(null, null, false)");
shouldThrow("document.implementation.createDocument(null, null, 0)");
shouldThrow("document.implementation.createDocument(null, null, '')");
shouldThrow("document.implementation.createDocument(null, null, {})");
</script>
</body>
</html>
