third_party/blink/web_tests/fast/dom/DOMImplementation/createDocument-doctype-err-expected.txt

PASS doctype is an instance of DocumentType
PASS document.implementation.createDocument(null, null, doctype) did not throw exception.
PASS document.implementation.createDocument(null, null, undefined) did not throw exception.
PASS document.implementation.createDocument(null, null, null) did not throw exception.
PASS document.implementation.createDocument(null, null, false) threw exception TypeError: Failed to execute 'createDocument' on 'DOMImplementation': parameter 3 is not of type 'DocumentType'..
PASS document.implementation.createDocument(null, null, 0) threw exception TypeError: Failed to execute 'createDocument' on 'DOMImplementation': parameter 3 is not of type 'DocumentType'..
PASS document.implementation.createDocument(null, null, '') threw exception TypeError: Failed to execute 'createDocument' on 'DOMImplementation': parameter 3 is not of type 'DocumentType'..
PASS document.implementation.createDocument(null, null, {}) threw exception TypeError: Failed to execute 'createDocument' on 'DOMImplementation': parameter 3 is not of type 'DocumentType'..
PASS successfullyParsed is true

TEST COMPLETE
